backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pillow==10.3.0
numpy==1.26.4
scikit-learn==1.5.0
xgboost==2.0.3
torch==2.3.0
torchvision==0.18.0
segmentation-models-pytorch==0.3.3
ultralytics==8.2.0
transformers==4.41.0
rasterio==1.3.10
geopandas==0.14.4
shapely==2.0.4
scipy==1.13.1
matplotlib==3.9.0
opencv-python==4.9.0.80
timm==1.0.3
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.1
